[ArrayPartition] split into two versions
